clean up collection editor navbar on mobile
